worker/requirements.txt
celery==5.3.4
openai==1.3.5
psycopg2-binary==2.9.9
redis==5.0.1